pyproject.toml
[project]
name = "teluka"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "pillow>=12.2.0",
    "polars>=1.39.3",
    "pydantic>=2.13.0",
    "python-dotenv>=1.2.2",
    "python-fasthtml>=0.13.3",
    "supabase>=2.28.3",
]
